# swagger.yaml
openapi: 3.0.0
paths:
...
  /bot-accesses:
    post:
      operationId: BotAccessesController_create
      summary: Создает доступ
      description: >-
        Создает запись о доступе пользователя к боту, используется при создании
        бота
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBotAccessDto'
      responses:
        '201':
          description: Запись о доступе создана.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotAccess'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
      tags: &ref_8
        - botAccesses
      security: &ref_9
        - bearer: []
    get:
      operationId: BotAccessesController_findAll
      summary: Найти все доступы
      parameters: []
      responses:
        '200':
          description: Все доступы найдены.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BotAccess'
        '403':
          description: Отказ в доступе
      tags: *ref_8
      security: *ref_9
  /bot-accesses/{id}:
    get:
      operationId: BotAccessesController_findOne
      summary: Найти доступ по ID
      description: ВАЖНО! Находит доступ по botAccessId (запись о связи botId и userId)
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор botAccessId
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Доступ найден.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotAccess'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_8
      security: *ref_9
    patch:
      operationId: BotAccessesController_update
      summary: Изменить уровень доступа
      description: >-
        Изменяет уровень существующего доступа. Должен передаваться полностью
        объект с теми доступами, которые должны быть у пользователя.
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор botAccessId
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBotAccessDto'
      responses:
        '200':
          description: Доступ изменен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotAccess'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_8
      security: *ref_9
    delete:
      operationId: BotAccessesController_delete
      summary: Удалить доступ
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор botAccessId
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Доступ удален.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotAccess'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_8
      security: *ref_9
  /bot-accesses/{userId}/{botId}:
    get:
      operationId: BotAccessesController_getPermission
      summary: Получить доступ
      description: Позвоялет проверить уровень доступа по botId и userId
      parameters:
        - name: botId
          required: true
          in: path
          description: Идентификатор бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
        - name: userId
          required: true
          in: path
          description: Идентификатор пользователя
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Информация о доступе по botId и userId получена.
          content:
            application/json:
              schema:
                type: Permission
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_8
      security: *ref_9
  /bot-accesses/{botId}:
    post:
      operationId: BotAccessesController_shareAccess
      summary: Поделиться доступом
      description: >-
        Позвоялет поделиться доступом по botId и создать новый доступ, если
        пользователь имеет полный уровень доступа к данному боту
      parameters:
        - name: botId
          required: true
          in: path
          description: Идентификатор бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareBotAccessDto'
      responses:
        '201':
          description: Новый доступ создан.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotAccess'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_8
      security: *ref_9
...
  /bots/{id}:
    delete:
      operationId: BotsController_remove
      summary: Удаление бота
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Бот удален
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_14
      security: *ref_15
    patch:
      operationId: BotsController_update
      summary: Смена имени бота
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                botName:
                  type: string
                  example: Салон красоты
      responses:
        '200':
          description: Имя бота обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_14
      security: *ref_15
    get:
      operationId: BotsController_findOne
      summary: Получить данные бота по Id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Информация о боте по Id получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_14
      security: *ref_15
...
  /botTemplates:
    get:
      operationId: BotTemplatesController_findAll
      summary: Список шаблонов ботов
      parameters: []
      responses:
        '200':
          description: Шаблоны ботов успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BotTemplate'
        '403':
          description: Отказ в доступе
      tags: &ref_16
        - botTemplates
      security: &ref_17
        - bearer: []
    post:
      operationId: BotTemplatesController_create
      summary: Создать шаблон бота
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBotTemplateDto'
      responses:
        '201':
          description: Шаблон бота успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotTemplate'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
      tags: *ref_16
      security: *ref_17
  /botTemplates/{id}:
    get:
      operationId: BotTemplatesController_findOne
      summary: Получить шаблон бота по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор шаблона бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Шаблон бота успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotTemplate'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_16
      security: *ref_17
    patch:
      operationId: BotTemplatesController_update
      summary: Изменить шаблон бота
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор шаблона бота
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBotTemplateDto'
      responses:
        '200':
          description: Шаблон бота успешно изменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotTemplate'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_16
      security: *ref_17
...
info:
  title: API BotKits
  description: Ручки для команды frontend
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateBotAccessDto:
      type: object
      properties:
        botId:
          type: string
          description: Id бота
          example: 64ff94ef12477f1d0934c614
        permission:
          type: object
          description: уровень доступа
          example: &ref_24
            voronki: viewer
      required:
        - botId
        - permission
    Profile:
      type: object
      properties:
        username:
          type: string
          example: Ivan Ivanov
        phone:
          type: string
          example: '+79501364578'
        avatar:
          type: string
          example: https://i.pravatar.cc/300
        balance:
          type: number
          example: 1400
        accounts:
          type: array
          items:
            type: string
      required:
        - username
        - phone
        - avatar
        - balance
        - accounts
    Messenger:
      type: object
      properties:
        name:
          type: string
          example: VK
        page:
          example: vk.com/club1245321223
          type: array
          items:
            type: string
        accessKey:
          type: string
          example: 1685494522:AAHzRs4YFqckLvBVARVoUL0c3B1GFqlDpo
        url:
          type: string
          example: some_url
      required:
        - name
        - page
        - accessKey
        - url
    Bot:
      type: object
      properties:
        icon:
          type: string
          example: >-
            https://cdn.icon-icons.com/icons2/1233/PNG/512/1492718766-vk_83600.png
        botName:
          type: string
          example: Бот Автоответчик
        profile:
          $ref: '#/components/schemas/Profile'
        messenger:
          example:
            name: VK
            page: vk.com/club1245321223
            accessKey: 1685494522:AAHzRs4YFqckLvBVARVoUL0c3B1GFqlDpo
            url: some_url
          allOf:
            - $ref: '#/components/schemas/Messenger'
        botSettings:
          type: object
      required:
        - icon
        - botName
        - profile
        - messenger
        - botSettings
    Permission:
      type: object
      properties:
        voronki:
          type: string
          example: editor
        newsletters:
          type: string
          example: editor
        lists:
          type: string
          example: editor
        statistic:
          type: string
          example: editor
        dialogs:
          type: string
          example: editor
        crm:
          type: string
          example: editor
        mini_landing:
          type: string
          example: editor
      required:
        - voronki
        - newsletters
        - lists
        - statistic
        - dialogs
        - crm
        - mini_landing
    BotAccess:
      type: object
      properties:
        _id:
          type: string
          example: 64ff89e7faea577804940275
        userId:
          example: 64ff89e7faea577804940275
          allOf:
            - $ref: '#/components/schemas/Profile'
        botId:
          example: 64ff89e7faea577804940275
          allOf:
            - $ref: '#/components/schemas/Bot'
        permission:
          example:
            voronki: editor
            newsletters: editor
            lists: editor
            statistic: editor
            dialogs: editor
            crm: editor
            mini_landing: editor
          allOf:
            - $ref: '#/components/schemas/Permission'
        createdAt:
          format: date-time
          type: string
          example: '2023-09-12T15:29:12.117Z'
        updatedAt:
          format: date-time
          type: string
          example: '2023-09-12T15:29:12.117Z'
      required:
        - _id
        - userId
        - botId
        - permission
        - createdAt
        - updatedAt
    UpdateBotAccessDto:
      type: object
      properties:
        permission:
          description: уровень доступа
          allOf:
            - $ref: '#/components/schemas/Permission'
      required:
        - permission
    ShareBotAccessDto:
      type: object
      properties:
        email:
          type: string
          description: Email зарегистрированного пользователя
          example: test@test.ru
        permission:
          type: object
          description: уровень доступа
          example: *ref_24
      required:
        - email
        - permission
...
    BotTemplate:
      type: object
      properties:
        icon:
          type: string
          example: >-
            https://cdn.icon-icons.com/icons2/1233/PNG/512/1492718766-vk_83600.png
        title:
          type: string
          example: Доставка еды
        description:
          type: string
          example: Бот для создания заказов
        features:
          example:
            - Создание заказов
            - Редактирование заказов
          type: array
          items:
            type: string
        settings:
          type: object
          example:
            Приветствие: Я бот для создания заказов
            Инлайн_кнопка: Текст кнопки
      required:
        - icon
        - title
        - description
        - features
        - settings
    CreateBotTemplateDto:
      type: object
      properties:
        icon:
          type: string
          example: >-
            https://cdn.icon-icons.com/icons2/1233/PNG/512/1492718766-vk_83600.png
        title:
          type: string
          example: Доставка еды
        description:
          type: string
          example: Бот для создания заказов
        features:
          example:
            - Создание заказов
            - Редактирование заказов
          type: array
          items:
            type: string
        settings:
          type: object
          example:
            Приветствие: Я бот для создания заказов
            Инлайн_кнопка: Текст кнопки
      required:
        - title
    UpdateBotTemplateDto:
      type: object
      properties:
        icon:
          type: string
          example: >-
            https://cdn.icon-icons.com/icons2/1233/PNG/512/1492718766-vk_83600.png
        title:
          type: string
          example: Доставка еды
        description:
          type: string
          example: Бот для создания заказов
        features:
          example:
            - Создание заказов
            - Редактирование заказов
          type: array
          items:
            type: string
        settings:
          type: object
          example:
            Приветствие: Я бот для создания заказов
            Инлайн_кнопка: Текст кнопки
...